Ранее написанный код разделен на модули
